GuessTheNumber fixes

+++ Lesson_1/GuessTheNumber.java
@@ -1,7 +1,7 @@
 public class GuessTheNumber {
 	public static void main(String[] args) {
-		short secretNumber = 33;
-		short userNumber = 0;
+		byte secretNumber = 33;
+		byte userNumber = 0;
 		while (userNumber != secretNumber) {
 			System.out.println("Введенное вами число больше (меньше) того, что загадал компьютер > " + userNumber);
 			userNumber++;
